+++ NEWS.md
@@ -11,6 +11,10 @@
 * `r2_nakagawa()` now returns r-squared for models with singular fit, where no random effect variances could be computed. The r-squared then does not take random effect variances into account. This behaviour was changed to be in line with `MuMIn::r.squaredGLMM()`, which returned a value for models with singular fit.
 * `check_distribution()` now detects negative binomial and zero-inflated distributions. Furthermore, attempt to improve accuracy.
 
+## New check-functions
+
+* `check_homogeneity()` to check models for homogeneity of variances.
+
 ## Bug fixes
 
 * Fix issues with `compare_performance()` and row-ordering.